+++ src/main/java/hu/multiteam/papyrum/block/PapyrumBlocks.java
@@ -2,11 +2,14 @@ package hu.multiteam.papyrum.block;
 
 import hu.multiteam.papyrum.Papyrum;
 import hu.multiteam.papyrum.item.PapyrumItems;
+import net.minecraft.core.BlockPos;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.RotatedPillarBlock;
 import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.MapColor;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.registries.DeferredRegister;
@@ -21,6 +24,7 @@ public class PapyrumBlocks {
     public static final RegistryObject<RotatedPillarBlock> CARDBOARD_BOX = register("cardboard_box", () -> new RotatedPillarBlock(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_BROWN)), new Item.Properties());
     public static final RegistryObject<Block> CARDBOARD_BLOCK = register("cardboard_block", () -> new Block(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_BROWN)), new Item.Properties());
     public static final RegistryObject<Block> CARDBOARD_TUBE = register("cardboard_tube", () -> new CardboardTubeBlock(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_BROWN)), new Item.Properties());
+    public static final RegistryObject<Block> CARDBOARD_WINDOW_BLOCK = register("cardboard_window_block", () -> new CardboardWindowBlock(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_BROWN).isViewBlocking(PapyrumBlocks::never)), new Item.Properties());
 
     public static void register(IEventBus bus){
         BLOCKS.register(bus);
@@ -32,5 +36,8 @@ public class PapyrumBlocks {
         RegistryObject<T> ret = registerNoItem(name, block);
         PapyrumItems.ITEMS.register(name, () -> new BlockItem(ret.get(), itemProperties));
         return ret;
+    }
+    private static boolean never(BlockState blockState, BlockGetter blockGetter, BlockPos blockPos) {
+        return false;
     }
 }