+++ changelog.md
@@ -6,3 +6,5 @@
 - added cardboard box block
 - added cardboard block
 - added recipe for the cardboard box and block
+- added cardboard tube block
+-